Internal Class for Slab and Shrinkage Models

This alias documents the S3 class savvySh_model, which is returned by the savvySh function. Instances of this class store the fitted shrinkage estimators, predicted values, model information, etc.

Details

Typically, you do not construct savvySh_model objects directly. Instead, they are produced when you run savvySh(...). Several S3 methods such as coef.savvySh_model, predict.savvySh_model, and summary.savvySh_model exist for extracting information from or summarizing a savvySh_model.