Extract Coefficients for a Slab and Shrinkage Model

Extracts the estimated regression coefficients from a savvySh_model object, specifying which shrinkage estimator to return, if no estimator specified, then return the fist avaliable estimator for the chosen model_class.

Usage

# S3 method for class 'savvySh_model'
coef(object, estimator = NULL, ...)

Arguments

object

A fitted model object of class savvySh_model (created by savvySh).

estimator

A character string naming which estimator's coefficients to extract. Examples include "St", "DSh", "Sh", "SR", "GSR", and so on. If NULL, the first available estimator is used.

...

Additional arguments passed to predict.savvySh_model.

Value

A named numeric vector of regression coefficients. If the model includes an intercept, it will appear as (Intercept) in the returned vector.

Details

This function calls predict.savvySh_model internally with type = "coefficients" to retrieve the requested coefficient vector. If an estimator is not supplied, it defaults to the first available estimator in the fitted model (for example, "St", "DSh", "SR", "GSR", etc.). If the requested estimator does not exist in the model object, an error is raised.

See also

Author

Ziwei Chen, Vali Asimit, Marina Anca Cidota, Jennifer Asimit
Maintainer: Ziwei Chen <ziwei.chen.3@citystgeorges.ac.uk>