Extract Coefficients From a Slab and Shrinkage Model Object

Extract the estimated coefficients from a fitted Slab and Shrinkage linear regression model object.

Usage

# S3 method for class 'savvySh_model'
coef(object, estimator = NULL, ...)

Arguments

object

A fitted model object of class savvySh_model.

estimator

The specific estimator to use for extracting coefficients. For example, "St", "DSh", etc. This must match one of the estimators in the fitted savvySh_model object. Defaults to the first available estimator if not specified.

...

Additional arguments passed to the predict method.

Value

A named vector of coefficients for the specified estimator.

See also

Author

Ziwei Chen, Vali Asimit, Marina Anca Cidota, Jennifer Asimit
Maintainer: Ziwei Chen <ziwei.chen.3@bayes.city.ac.uk>