Predict for Slab and Shrinkage Linear Regression Models

This function predicts fitted values or coefficients for a fitted Slab and Shrinkage linear regression model created using the savvySh function.

Usage

# S3 method for class 'savvySh_model'
predict(
  object,
  newx = NULL,
  type = c("response", "coefficients"),
  estimator = NULL,
  ...
)

Arguments

object

Fitted "savvySh_model" object returned by savvySh.

newx

Matrix of new data for which predictions are to be made. Must have the same number of columns as the training data. This argument is required for type = "response".

type

Type of prediction required. Can be "response" (fitted values) or "coefficients". Defaults to "response".

estimator

The specific estimator to use for prediction. For example, "St", "DSh" if model_class = "Multiplicative", etc. This must match one of the estimators in the fitted savvySh_model object. Defaults to the first available estimator if not specified.

...

Additional arguments.

Value

Predicted values for type = "response" or coefficients for type = "coefficients".

See also

Author

Ziwei Chen, Vali Asimit, Marina Anca Cidota, Jennifer Asimit
Maintainer: Ziwei Chen <ziwei.chen.3@citystgeorges.ac.uk>